requirements.txt
seaborn
pytorch-lightning
torchmetrics
torch
rich
lightning-bolts
hydra-core